add authentication

+++ app/Http/Controllers/MyProgramsController.php
@@ -5,27 +5,29 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\MyProgram;
 
-class MyProgramsController extends Controller
-{
-    function create($program_id, $user_id){
-      $relation = Myprogram::where([
-        ['program_id', '=', $program_id],
-        ['user_id', '=', $user_id],
-      ])->first();
+class MyProgramsController extends Controller {
 
-      if ($relation == null) {
-        $relation = new MyProgram;
-        $relation->program_id = $program_id;
-        $relation->user_id = $user_id;
-        $relation->save();
+    function create($program_id, $user_id) {
+        if (Auth::check()) {
+            $relation = Myprogram::where([
+                            ['program_id', '=', $program_id],
+                            ['user_id', '=', $user_id],
+                    ])->first();
 
-        return redirect()->back()->with('alert', 'Participate successfully');
-      }
-      else
-      {
-        $relation->delete();
+            if ($relation == null) {
+                $relation = new MyProgram;
+                $relation->program_id = $program_id;
+                $relation->user_id = $user_id;
+                $relation->save();
 
-        return redirect()->back()->with('alert', 'Cancel participating successfully');
-      }
+                return redirect()->back()->with('alert', 'Participate successfully');
+            } else {
+                $relation->delete();
+
+                return redirect()->back()->with('alert', 'Cancel participating successfully');
+            }
+        }
+        return redirect('');
     }
+
 }
